+++ lib/ng-vcl/src/alert/CHANGELOG.md
@@ -3,3 +3,4 @@
 ### BREAKING CHANGES
 - Cancel/Close emits a value instead of throwing an error
 - AlertResult now contains an `action` property showing the type `confirm`, `cancel` or `close`
+- Removed some barely used options